Refractored Waypoint ID to be a long now instead of a int to follow backend. Implemented executeJob functionality. Rebuild release.

+++ SmartRacecar/src/main/java/be/uantwerpen/fti/ds/sc/smartracecar/common/WayPoint.java
@@ -3,16 +3,16 @@ package be.uantwerpen.fti.ds.sc.smartracecar.common;
 //Extended common of a Point which is used for waypoints. Supers the coordinates from the Point class.
 public class WayPoint extends Point {
 
-    private int ID = 0; // waypoint ID given by RaceCarManager.
+    private long ID = 0; // waypoint ID given by RaceCarManager.
     private int weight = 0; //weight used for pathfinding algorthims.
 
-    WayPoint(int ID,float x, float y,float z, float w, int weight) {
+    WayPoint(long ID,float x, float y,float z, float w, int weight) {
         super(x, y, z, w);
         this.ID = ID;
         this.weight = weight;
     }
 
-    public int getID() {
+    public long getID() {
         return ID;
     }
 }
